+++ Semana 2/node/src/model/Appointments.ts
@@ -1,5 +1,15 @@
 import { startOfHour, parseISO } from "date-fns";
-import { Entity, Column, PrimaryGeneratedColumn } from "typeorm";
+import {
+  Entity,
+  Column,
+  PrimaryGeneratedColumn,
+  CreateDateColumn,
+  UpdateDateColumn,
+  ManyToOne,
+  JoinColumn,
+} from "typeorm";
+
+import User from "./Users";
 
 // A model é responsavel pela estrutura do dado que a aplicacao utiliza
 @Entity("appointments")
@@ -8,10 +18,20 @@ class Appointments {
   id: string;
 
   @Column()
-  provider: string;
+  provider_id: string;
 
-  @Column('timestamp with time zone')
+  @ManyToOne(() => User) // Relacao Muitos para UM
+  @JoinColumn({name: 'provider_id'}) // Relacao Muitos para UM
+  provider: User;
+
+  @Column("timestamp with time zone")
   date: Date;
+
+  @CreateDateColumn()
+  created_at: Date;
+
+  @UpdateDateColumn()
+  updated_at: Date;
 
   // Com o typeorm nao é necessario criar um construtor, porem para parar de dar erro é necessario desabilitar no tsconfig a propriedade strictPropertyInitialization
 }